python/packages/stoa-recalld/pyproject.toml
[project]
name = "stoa-recalld"
version = "0.1.0"
description = "Stoa long-lived recall daemon — MemPalace-backed Unix-socket server."
requires-python = ">=3.13"
license = "MIT"
readme = "README.md"
dependencies = [
    "mempalace>=3.3.5,<4",
    "pyyaml>=6.0",
]

[project.scripts]
stoa-recalld = "stoa_recalld.cli:main"

[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stoa_recalld"]

[tool.basedpyright]
include = ["src/stoa_recalld", "tests"]
exclude = ["**/__pycache__", "build", "dist"]
pythonVersion = "3.13"
typeCheckingMode = "standard"
